Move serialization utils to the debugger
