+++ .changeset/empty-carrots-attack.md
@@ -1,0 +1,6 @@
+---
+'@solid-devtools/debugger': minor
+'@solid-devtools/shared': minor
+---
+
+Move value serialization logic to debugger
