+++ packages/debugger/src/inspector.ts
@@ -1,11 +1,15 @@
-import { Mapped, NodeID, NodeType } from '@solid-devtools/shared/graph'
-import { ElementMap, encodeValue, ValueType } from '@solid-devtools/shared/serialize'
-import { pushToArrayProp } from '@solid-devtools/shared/utils'
-import { $PROXY, untrack } from 'solid-js'
+import { $PROXY, Accessor, createEffect, untrack } from 'solid-js'
 import { DEV as STORE_DEV } from 'solid-js/store'
+import { throttle } from '@solid-primitives/scheduled'
+import { Mapped, NodeID, NodeType, EncodedValue, ValueType } from '@solid-devtools/shared/graph'
+import { untrackedCallback } from '@solid-devtools/shared/primitives'
+import { pushToArrayProp } from '@solid-devtools/shared/utils'
+import { DebuggerEventHub } from './plugin'
+import { walkSolidRoot } from './roots'
 import { Core, Solid, ValueUpdateListener } from './types'
-import { observeValueUpdate, removeValueUpdateObserver } from './update'
+import { makeSolidUpdateListener, observeValueUpdate, removeValueUpdateObserver } from './update'
 import {
+  createBatchedUpdateEmitter,
   getComponentRefreshNode,
   getDisplayName,
   getNodeName,
@@ -19,6 +23,7 @@ import {
   markOwnerName,
   markOwnerType,
 } from './utils'
+import { ElementMap, encodeValue } from './serialize'
 
 const DEV = STORE_DEV!
 
@@ -41,13 +46,11 @@ function mapSignalNode(
 
   // Check if is a store
   if (isSolidStore(node)) {
-    console.log('store', node)
     return {
       type: NodeType.Store,
       id,
       name: getDisplayName((value as any)[DEV.$NAME]),
-      // value: encodeValue(value, false, $elementMap),
-      value: { type: ValueType.Object, value: 0 },
+      value: encodeValue(value, false, $elementMap),
       // TODO: top-level values can be observed too, it's the "_" property
       observers: [],
     }
@@ -241,5 +244,141 @@ export function inspectStoreNode(
     else node[DEV.$ON_UPDATE] = node[DEV.$ON_UPDATE]!.filter(h => h.symbol !== symbol)
     set.delete(node)
     forEachStoreProp(node, (_, child) => set.has(child) && untrackStore(child))
+  }
+}
+
+/**
+ * Plugin module
+ */
+export function createInspector({
+  debuggerEnabled,
+  eventHub,
+}: {
+  debuggerEnabled: Accessor<boolean>
+  eventHub: DebuggerEventHub
+}) {
+  const inspected = {
+    elementMap: new ElementMap(),
+    signalMap: {} as SignalMap,
+    owner: null as Solid.Owner | null,
+    signals: new Set<NodeID>(),
+    props: new Set<string>(),
+    value: false,
+    getValue: () => null as unknown,
+  }
+
+  const getElementById = (id: NodeID): HTMLElement | undefined => inspected.elementMap.get(id)
+
+  const pushSignalUpdate = createBatchedUpdateEmitter<{
+    id: NodeID
+    value: EncodedValue<boolean>
+  }>(updates => eventHub.emit('SignalUpdates', updates))
+  const onSignalUpdate: SignalUpdateHandler = untrackedCallback((id, value) => {
+    if (!debuggerEnabled() || !inspected.owner) return
+    const isSelected = inspected.signals.has(id)
+    pushSignalUpdate({ id, value: encodeValue(value, isSelected, inspected.elementMap) })
+  })
+
+  const triggerValueUpdate = (() => {
+    let updateNext = false
+    const forceValueUpdate = () => {
+      if (!debuggerEnabled() || !inspected.owner) return (updateNext = false)
+      eventHub.emit('ValueUpdate', {
+        value: encodeValue(inspected.getValue(), inspected.value, inspected.elementMap),
+        update: updateNext,
+      })
+      updateNext = false
+    }
+    const triggerValueUpdate = throttle(forceValueUpdate)
+    function handleValueUpdate(update: boolean, force = false) {
+      if (update) updateNext = true
+      if (force) forceValueUpdate()
+      else triggerValueUpdate()
+    }
+    return handleValueUpdate
+  })()
+
+  const setInspectedDetails = untrackedCallback((owner: Solid.Owner) => {
+    inspected.owner && clearOwnerObservers(inspected.owner)
+    inspected.props.clear()
+    inspected.signals.clear()
+    inspected.owner = owner
+    inspected.value = false
+    const result = collectOwnerDetails(owner, {
+      onSignalUpdate,
+      onValueUpdate: () => triggerValueUpdate(true),
+    })
+    eventHub.emit('InspectedNodeDetails', result.details)
+    inspected.signalMap = result.signalMap
+    inspected.elementMap = result.elementMap
+    inspected.getValue = result.getOwnerValue
+  })
+  const clearInspectedDetails = () => {
+    inspected.owner && clearOwnerObservers(inspected.owner)
+    inspected.owner = null
+    inspected.signals.clear()
+    inspected.props.clear()
+    inspected.value = false
+  }
+
+  function updateInspectedProps() {
+    if (!inspected.owner) return
+    const props = encodeComponentProps(inspected.owner, {
+      inspectedProps: inspected.props,
+      elementMap: inspected.elementMap,
+    })
+    props && eventHub.emit('PropsUpdate', props)
+  }
+
+  createEffect(() => {
+    // make sure we clear the owner observers when the plugin is disabled
+    if (!debuggerEnabled()) inspected.owner && clearOwnerObservers(inspected.owner)
+    // re-observe the owner when the plugin is enabled
+    else inspected.owner && setInspectedDetails(inspected.owner)
+
+    // update the owner details whenever there is a change in solid's internals
+    makeSolidUpdateListener(
+      throttle(() => {
+        updateInspectedProps()
+        triggerValueUpdate(false)
+      }, 150),
+    )
+  })
+
+  function setInspectedNode(data: { rootId: NodeID; nodeId: NodeID } | null) {
+    if (!data) return clearInspectedDetails()
+    const { rootId, nodeId } = data
+
+    const walkResult = walkSolidRoot(rootId, nodeId)
+    if (!walkResult || !walkResult.inspectedOwner) return clearInspectedDetails()
+
+    setInspectedDetails(walkResult.inspectedOwner)
+  }
+
+  // TODO: this shouldn't return the value, but send ot though the event hub
+  function setInspectedSignal(id: NodeID, selected: boolean): EncodedValue<boolean> | null {
+    const signal = inspected.signalMap[id] as Solid.Signal | undefined
+    if (!signal) return null
+    if (selected) inspected.signals.add(id)
+    else inspected.signals.delete(id)
+    return untrack(() => encodeValue(signal.value, selected, inspected.elementMap))
+  }
+  function setInspectedProp(key: NodeID, selected: boolean) {
+    if (selected) inspected.props.add(key)
+    else inspected.props.delete(key)
+    updateInspectedProps()
+  }
+  function setInspectedValue(selected: boolean) {
+    if (!inspected.owner) return null
+    inspected.value = selected
+    triggerValueUpdate(false, true)
+  }
+
+  return {
+    setInspectedNode,
+    setInspectedSignal,
+    setInspectedProp,
+    setInspectedValue,
+    getElementById,
   }
 }
