+++ packages/frontend/src/modules/inspector/index.ts
@@ -2,8 +2,7 @@ import { batch, createEffect, createSelector, createSignal } from 'solid-js'
 import { createStore, produce } from 'solid-js/store'
 import { Writable } from 'type-fest'
 import { createSimpleEmitter } from '@solid-primitives/event-bus'
-import { Mapped, NodeID, NodeType } from '@solid-devtools/shared/graph'
-import { EncodedValue } from '@solid-devtools/shared/serialize'
+import { Mapped, NodeID, NodeType, EncodedValue } from '@solid-devtools/shared/graph'
 import { defer, untrackedCallback } from '@solid-devtools/shared/primitives'
 import type { Structure } from '../structure'
