+++ packages/shared/src/index.ts
@@ -1,6 +1,5 @@
 export * as bridge from './bridge'
 export * as graph from './graph'
 export * as primitives from './primitives'
-export * as serialize from './serialize'
 export * as utils from './utils'
 export * as variables from './variables'
