+++ packages/shared/src/variables.ts
@@ -1,5 +1,3 @@
 export const WINDOW_PROJECTPATH_PROPERTY = '$sdt_projectPath'
 export const WINDOW_WRAP_STORE_PROPERTY = '$sdt_wrapStore'
 export const LOCATION_ATTRIBUTE_NAME = 'data-source-loc'
-export const UNNAMED = '(unnamed)'
-export const INTERNAL = Symbol('internal')
